Add new commands to the package.json file
Now there are to options when compiling the styles,
you can do it with the tailwind command which will
use the tailwind-cli for compiling the stylesheets,
or, you can use the postcss command which do almost
the same as tailwind but with the postcss-cli.

To have to opportunity of using the postcss-cli on
the project, the postcss-cli package was added as a
dev dependency.